Arreglado lo de que no guardaba

+++ app/src/main/java/com/example/inventory/ui/home/HomeScreen.kt
@@ -183,7 +183,7 @@ private fun InventoryItem(
                 )
             }
             Text(
-                text = stringResource(R.string.in_stock, item.rating),
+                text = stringResource(R.string.in_stock, item.rating.toInt()),
                 style = MaterialTheme.typography.titleMedium
             )
         }
